## Optimistic concurrency control. Hyperspace doesn't include pessimistic concurrency control because it doesn't need it. Immutable entities never change. Version entities are always a new version.
Optimistic concurrency control uses the last `AsAt` timestamp retrieved from hiperspace And checks this value hasn't changed when a new version is added.
Concurrent updates are blocked because other parts of the Element may have changed and these would be overwritten by the change.

## %sql
`System.ComponentModel.Annotations` is now part of the framework install, so has been replaced with check for 'Microsoft.EntityFrameworkCore` for auto-add of reference

## %odata
Include mapping for odata export of values as json text for editing within a client

## Event Forwarding
Forward events in SubSpace SetSpaces to any listeners added to the equvilent SetSpaces in the provider hiperspace.  This enables seamless capture of favct table events for continueous agggregation