https://www.cepheis.com/hiperspace/20251023 ## Overview
This release extends the **message** functionality to behave more like a stored procedure call,
binding all *entity/segment/aspect/view* results to the local subspace when it completes.

Initially the `RemoteLabel`
flag was used to indicate which tier should invoke the message, and return results.  Test with **Hiperspace.DB**
highlighted that the flag is not needed - the message is now executed by the lowest *Hiperspace* that can provide a value.